@@ -10,8 +10,14 @@
     "dev": "node_modules/.bin/webpack-dev-server"
   },
   "devDependencies": {
+    "babel-core": "^6.3.17",
     "babel-eslint": "^3.1.8",
-    "babel-loader": "^5.1.3",
+    "babel-loader": "^6.2.0",
+    "babel-plugin-syntax-async-functions": "^6.3.13",
+    "babel-plugin-transform-object-rest-spread": "^6.3.13",
+    "babel-plugin-transform-regenerator": "^6.3.13",
+    "babel-preset-es2015": "^6.3.13",
+    "babel-preset-react": "^6.3.13",
     "css-loader": "^0.13.1",
     "eslint": "^0.21.2",
     "eslint-plugin-react": "^2.3.0",
@@ -26,13 +32,15 @@
     "less": "^2.5.1",
     "less-loader": "^2.2.0",
     "load-grunt-tasks": "^3.2.0",
+    "minimatch": "^3.0.0",
     "style-loader": "^0.12.2",
-    "webpack": "^1.7.3",
-    "webpack-dev-server": "^1.7.0"
+    "webpack": "^1.12.9",
+    "webpack-dev-server": "^1.14.0"
   },
   "dependencies": {
-    "babel-runtime": "^5.4.7",
+    "babel-polyfill": "^6.3.14",
     "bootstrap": "^3.3.4",
+    "history": "^1.13.1",
     "immutable": "^3.7.3",
     "lodash": "^3.9.1",
     "react": "^0.14.3",
@@ -41,7 +49,7 @@
     "react-hotkeys": "git://github.com/thomasboyt/react-hotkeys#fix-map-change",
     "react-immutable-render-mixin": "^0.8.1",
     "react-redux": "^4.0.0",
-    "react-router": "^0.13.5",
+    "react-router": "^1.0.2",
     "redux": "^3.0.4",
     "redux-thunk": "^1.0.0"
   }
